+++ test/template.js
@@ -3,11 +3,11 @@ module('Template');
 test("initialize", function () {
     var tpl = new yamvc.view.Template({
         config : {
-            id : 'tpl-test'
-        },
-        html : [
-            '{{Test}}'
-        ]
+            id : 'tpl-test',
+            tpl : [
+                '{{Test}}'
+            ]
+        }
     });
 
     ok(tpl instanceof yamvc.view.Template);